Add getLocation method to Space API
The location of the space needs to be accessible (but not mutable) in
the Player API. The reason for this is because the player can pass the
space to other senses, that then use its location. This solution isn't
perfect, but it doesn't hurt. Should revisit if more things "need" to
be added to the Player API in the future.