Introduce sensed space and sensed unit entities (#129)
Whenever a unit gets a space after sensing an area, that space will be
a `SensedSpace` and the unit on it (if any) a `SensedUnit`. These new
entities represent the spaces and units as they were sensed by the
unit. Methods called on these objects are now relative to the unit that
sensed them:

 * `space.getLocation()`: Returns the relative location of the space
   from the location of the unit that sensed it.
 
 * `unit.isEnemy()`: Returns whether the unit is considered an enemy by
   the unit that sensed it.

This allowed the Unit API to be simplified, removing these methods:

 * `unit.isWarrior()` and `unit.isPlayer()`: Units that are enemies to the
   warrior will now consider the warrior an enemy.

Also, `unit.isHostile()` was renamed to `unit.isEnemy()` and
`unit.isFriendly()` was removed.